remove LoadObject from PlannedBucketController
